+++ src/pyodide/worker/worker-dist.js
@@ -112,6 +112,12 @@ var worker_config_PYODIDE_WORKER_CONFIG = {
 function _typeof(o) { "@babel/helpers - typeof"; return _typeof = "function" == typeof Symbol && "symbol" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && "function" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? "symbol" : typeof o; }, _typeof(o); }
 function _regenerator() { /*! regenerator-runtime -- Copyright (c) 2014-present, Facebook, Inc. -- license (MIT): https://github.com/babel/babel/blob/main/packages/babel-helpers/LICENSE */ var e, t, r = "function" == typeof Symbol ? Symbol : {}, n = r.iterator || "@@iterator", o = r.toStringTag || "@@toStringTag"; function i(r, n, o, i) { var c = n && n.prototype instanceof Generator ? n : Generator, u = Object.create(c.prototype); return _regeneratorDefine2(u, "_invoke", function (r, n, o) { var i, c, u, f = 0, p = o || [], y = !1, G = { p: 0, n: 0, v: e, a: d, f: d.bind(e, 4), d: function d(t, r) { return i = t, c = 0, u = e, G.n = r, a; } }; function d(r, n) { for (c = r, u = n, t = 0; !y && f && !o && t < p.length; t++) { var o, i = p[t], d = G.p, l = i[2]; r > 3 ? (o = l === n) && (u = i[(c = i[4]) ? 5 : (c = 3, 3)], i[4] = i[5] = e) : i[0] <= d && ((o = r < 2 && d < i[1]) ? (c = 0, G.v = n, G.n = i[1]) : d < l && (o = r < 3 || i[0] > n || n > l) && (i[4] = r, i[5] = n, G.n = l, c = 0)); } if (o || r > 1) return a; throw y = !0, n; } return function (o, p, l) { if (f > 1) throw TypeError("Generator is already running"); for (y && 1 === p && d(p, l), c = p, u = l; (t = c < 2 ? e : u) || !y;) { i || (c ? c < 3 ? (c > 1 && (G.n = -1), d(c, u)) : G.n = u : G.v = u); try { if (f = 2, i) { if (c || (o = "next"), t = i[o]) { if (!(t = t.call(i, u))) throw TypeError("iterator result is not an object"); if (!t.done) return t; u = t.value, c < 2 && (c = 0); } else 1 === c && (t = i.return) && t.call(i), c < 2 && (u = TypeError("The iterator does not provide a '" + o + "' method"), c = 1); i = e; } else if ((t = (y = G.n < 0) ? u : r.call(n, G)) !== a) break; } catch (t) { i = e, c = 1, u = t; } finally { f = 1; } } return { value: t, done: y }; }; }(r, o, i), !0), u; } var a = {}; function Generator() {} function GeneratorFunction() {} function GeneratorFunctionPrototype() {} t = Object.getPrototypeOf; var c = [][n] ? t(t([][n]())) : (_regeneratorDefine2(t = {}, n, function () { return this; }), t), u = GeneratorFunctionPrototype.prototype = Generator.prototype = Object.create(c); function f(e) { return Object.setPrototypeOf ? Object.setPrototypeOf(e, GeneratorFunctionPrototype) : (e.__proto__ = GeneratorFunctionPrototype, _regeneratorDefine2(e, o, "GeneratorFunction")), e.prototype = Object.create(u), e; } return GeneratorFunction.prototype = GeneratorFunctionPrototype, _regeneratorDefine2(u, "constructor", GeneratorFunctionPrototype), _regeneratorDefine2(GeneratorFunctionPrototype, "constructor", GeneratorFunction), GeneratorFunction.displayName = "GeneratorFunction", _regeneratorDefine2(GeneratorFunctionPrototype, o, "GeneratorFunction"), _regeneratorDefine2(u), _regeneratorDefine2(u, o, "Generator"), _regeneratorDefine2(u, n, function () { return this; }), _regeneratorDefine2(u, "toString", function () { return "[object Generator]"; }), (_regenerator = function _regenerator() { return { w: i, m: f }; })(); }
 function _regeneratorDefine2(e, r, n, t) { var i = Object.defineProperty; try { i({}, "", {}); } catch (e) { i = 0; } _regeneratorDefine2 = function _regeneratorDefine(e, r, n, t) { if (r) i ? i(e, r, { value: n, enumerable: !t, configurable: !t, writable: !t }) : e[r] = n;else { var o = function o(r, n) { _regeneratorDefine2(e, r, function (e) { return this._invoke(r, n, e); }); }; o("next", 0), o("throw", 1), o("return", 2); } }, _regeneratorDefine2(e, r, n, t); }
+function _slicedToArray(r, e) { return _arrayWithHoles(r) || _iterableToArrayLimit(r, e) || _unsupportedIterableToArray(r, e) || _nonIterableRest(); }
+function _nonIterableRest() { throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }
+function _unsupportedIterableToArray(r, a) { if (r) { if ("string" == typeof r) return _arrayLikeToArray(r, a); var t = {}.toString.call(r).slice(8, -1); return "Object" === t && r.constructor && (t = r.constructor.name), "Map" === t || "Set" === t ? Array.from(r) : "Arguments" === t || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(t) ? _arrayLikeToArray(r, a) : void 0; } }
+function _arrayLikeToArray(r, a) { (null == a || a > r.length) && (a = r.length); for (var e = 0, n = Array(a); e < a; e++) n[e] = r[e]; return n; }
+function _iterableToArrayLimit(r, l) { var t = null == r ? null : "undefined" != typeof Symbol && r[Symbol.iterator] || r["@@iterator"]; if (null != t) { var e, n, i, u, a = [], f = !0, o = !1; try { if (i = (t = t.call(r)).next, 0 === l) { if (Object(t) !== t) return; f = !1; } else for (; !(f = (e = i.call(t)).done) && (a.push(e.value), a.length !== l); f = !0); } catch (r) { o = !0, n = r; } finally { try { if (!f && null != t.return && (u = t.return(), Object(u) !== u)) return; } finally { if (o) throw n; } } return a; } }
+function _arrayWithHoles(r) { if (Array.isArray(r)) return r; }
 function ownKeys(e, r) { var t = Object.keys(e); if (Object.getOwnPropertySymbols) { var o = Object.getOwnPropertySymbols(e); r && (o = o.filter(function (r) { return Object.getOwnPropertyDescriptor(e, r).enumerable; })), t.push.apply(t, o); } return t; }
 function _objectSpread(e) { for (var r = 1; r < arguments.length; r++) { var t = null != arguments[r] ? arguments[r] : {}; r % 2 ? ownKeys(Object(t), !0).forEach(function (r) { _defineProperty(e, r, t[r]); }) : Object.getOwnPropertyDescriptors ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(t)) : ownKeys(Object(t)).forEach(function (r) { Object.defineProperty(e, r, Object.getOwnPropertyDescriptor(t, r)); }); } return e; }
 function _defineProperty(e, r, t) { return (r = _toPropertyKey(r)) in e ? Object.defineProperty(e, r, { value: t, enumerable: !0, configurable: !0, writable: !0 }) : e[r] = t, e; }
@@ -148,7 +154,7 @@ function handleExecute(_x, _x2) {
  */
 function _handleExecute() {
   _handleExecute = _asyncToGenerator(/*#__PURE__*/_regenerator().m(function _callee(data, workerState) {
-    var code, filename, namespace, start, stdout, stderr, missive, figures, error, result, _captureOutputs, _captureOutputs2, _t;
+    var code, filename, namespace, start, stdout, stderr, missive, figures, error, result, originalValues, keysToRestore, _i, _Object$entries, _Object$entries$_i, key, value, _i2, _keysToRestore, _key, _captureOutputs, _captureOutputs2, _t;
     return _regenerator().w(function (_context) {
       while (1) switch (_context.n) {
         case 0:
@@ -178,44 +184,74 @@ function _handleExecute() {
 
           // Execute with or without namespace
           if (!(namespace !== undefined)) {
-            _context.n = 6;
-            break;
-          }
-          console.log(worker_config_PYODIDE_WORKER_CONFIG.MESSAGES.EXEC_NAMESPACE);
-          if (!result.needsAsync) {
-            _context.n = 4;
-            break;
-          }
-          console.log("🔧 [Worker] Running async with namespace");
-          _context.n = 3;
-          return workerState.pyodide.runPythonAsync(result.code, {
-            globals: workerState.pyodide.toPy(namespace)
-          });
-        case 3:
-          _context.n = 5;
-          break;
-        case 4:
-          workerState.pyodide.runPython(result.code, {
-            globals: workerState.pyodide.toPy(namespace)
-          });
-        case 5:
-          _context.n = 9;
-          break;
-        case 6:
-          console.log(worker_config_PYODIDE_WORKER_CONFIG.MESSAGES.EXEC_GLOBAL);
-          if (!result.needsAsync) {
             _context.n = 8;
             break;
           }
-          console.log("🔧 [Worker] Running async in global scope");
-          _context.n = 7;
+          console.log(worker_config_PYODIDE_WORKER_CONFIG.MESSAGES.EXEC_NAMESPACE);
+          console.log("🔧 [Worker] Namespace variables:", Object.keys(namespace));
+
+          // Store original values to restore later
+          originalValues = {};
+          keysToRestore = []; // Temporarily set namespace variables
+          for (_i = 0, _Object$entries = Object.entries(namespace); _i < _Object$entries.length; _i++) {
+            _Object$entries$_i = _slicedToArray(_Object$entries[_i], 2), key = _Object$entries$_i[0], value = _Object$entries$_i[1];
+            // Store original value if it exists
+            if (workerState.pyodide.globals.has(key)) {
+              originalValues[key] = workerState.pyodide.globals.get(key);
+              keysToRestore.push(key);
+            } else {
+              keysToRestore.push(key);
+            }
+            workerState.pyodide.globals.set(key, value);
+            console.log("\uD83D\uDD27 [Worker] Set ".concat(key, " = ").concat(value));
+          }
+          _context.p = 3;
+          if (!result.needsAsync) {
+            _context.n = 5;
+            break;
+          }
+          console.log("🔧 [Worker] Running async with namespace");
+          _context.n = 4;
           return workerState.pyodide.runPythonAsync(result.code);
+        case 4:
+          _context.n = 6;
+          break;
+        case 5:
+          workerState.pyodide.runPython(result.code);
+        case 6:
+          _context.p = 6;
+          // Clean up namespace variables
+          for (_i2 = 0, _keysToRestore = keysToRestore; _i2 < _keysToRestore.length; _i2++) {
+            _key = _keysToRestore[_i2];
+            if (originalValues.hasOwnProperty(_key)) {
+              // Restore original value
+              workerState.pyodide.globals.set(_key, originalValues[_key]);
+              console.log("\uD83D\uDD27 [Worker] Restored ".concat(_key, " to original value"));
+            } else {
+              // Delete the variable we added
+              workerState.pyodide.globals.delete(_key);
+              console.log("\uD83D\uDD27 [Worker] Removed ".concat(_key, " from globals"));
+            }
+          }
+          return _context.f(6);
         case 7:
-          _context.n = 9;
+          _context.n = 11;
           break;
         case 8:
-          workerState.pyodide.runPython(result.code);
+          console.log(worker_config_PYODIDE_WORKER_CONFIG.MESSAGES.EXEC_GLOBAL);
+          if (!result.needsAsync) {
+            _context.n = 10;
+            break;
+          }
+          console.log("🔧 [Worker] Running async in global scope");
+          _context.n = 9;
+          return workerState.pyodide.runPythonAsync(result.code);
         case 9:
+          _context.n = 11;
+          break;
+        case 10:
+          workerState.pyodide.runPython(result.code);
+        case 11:
           console.log("🔧 [Worker] Execution completed, capturing outputs");
           _captureOutputs = captureOutputs(workerState.pyodide);
           stdout = _captureOutputs.stdout;
@@ -223,10 +259,10 @@ function _handleExecute() {
           missive = _captureOutputs.missive;
           figures = _captureOutputs.figures;
           console.log("🔧 [Worker] Captured outputs - stdout:", stdout.length, "stderr:", stderr.length, "missive:", missive, "figures:", figures.length);
-          _context.n = 11;
+          _context.n = 13;
           break;
-        case 10:
-          _context.p = 10;
+        case 12:
+          _context.p = 12;
           _t = _context.v;
           console.error("🔧 [Worker] Execution error:", _t);
           error = {
@@ -237,7 +273,7 @@ function _handleExecute() {
           stdout = _captureOutputs2.stdout;
           stderr = _captureOutputs2.stderr;
           figures = _captureOutputs2.figures;
-        case 11:
+        case 13:
           console.log("🔧 [Worker] Posting result");
           postResult({
             filename: filename,
@@ -249,10 +285,10 @@ function _handleExecute() {
             time: Date.now() - start,
             executedWithNamespace: namespace !== undefined
           });
-        case 12:
+        case 14:
           return _context.a(2);
       }
-    }, _callee, null, [[2, 10]]);
+    }, _callee, null, [[3,, 6, 7], [2, 12]]);
   }));
   return _handleExecute.apply(this, arguments);
 }
@@ -725,13 +761,13 @@ function file_loader_regeneratorDefine2(e, r, n, t) { var i = Object.definePrope
 function file_loader_asyncGeneratorStep(n, t, e, r, o, a, c) { try { var i = n[a](c), u = i.value; } catch (n) { return void e(n); } i.done ? t(u) : Promise.resolve(u).then(r, o); }
 function file_loader_asyncToGenerator(n) { return function () { var t = this, e = arguments; return new Promise(function (r, o) { var a = n.apply(t, e); function _next(n) { file_loader_asyncGeneratorStep(a, r, o, _next, _throw, "next", n); } function _throw(n) { file_loader_asyncGeneratorStep(a, r, o, _next, _throw, "throw", n); } _next(void 0); }); }; }
 function file_loader_typeof(o) { "@babel/helpers - typeof"; return file_loader_typeof = "function" == typeof Symbol && "symbol" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && "function" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? "symbol" : typeof o; }, file_loader_typeof(o); }
-function _slicedToArray(r, e) { return _arrayWithHoles(r) || _iterableToArrayLimit(r, e) || _unsupportedIterableToArray(r, e) || _nonIterableRest(); }
-function _nonIterableRest() { throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }
-function _iterableToArrayLimit(r, l) { var t = null == r ? null : "undefined" != typeof Symbol && r[Symbol.iterator] || r["@@iterator"]; if (null != t) { var e, n, i, u, a = [], f = !0, o = !1; try { if (i = (t = t.call(r)).next, 0 === l) { if (Object(t) !== t) return; f = !1; } else for (; !(f = (e = i.call(t)).done) && (a.push(e.value), a.length !== l); f = !0); } catch (r) { o = !0, n = r; } finally { try { if (!f && null != t.return && (u = t.return(), Object(u) !== u)) return; } finally { if (o) throw n; } } return a; } }
-function _arrayWithHoles(r) { if (Array.isArray(r)) return r; }
-function _createForOfIteratorHelper(r, e) { var t = "undefined" != typeof Symbol && r[Symbol.iterator] || r["@@iterator"]; if (!t) { if (Array.isArray(r) || (t = _unsupportedIterableToArray(r)) || e && r && "number" == typeof r.length) { t && (r = t); var _n = 0, F = function F() {}; return { s: F, n: function n() { return _n >= r.length ? { done: !0 } : { done: !1, value: r[_n++] }; }, e: function e(r) { throw r; }, f: F }; } throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); } var o, a = !0, u = !1; return { s: function s() { t = t.call(r); }, n: function n() { var r = t.next(); return a = r.done, r; }, e: function e(r) { u = !0, o = r; }, f: function f() { try { a || null == t.return || t.return(); } finally { if (u) throw o; } } }; }
-function _unsupportedIterableToArray(r, a) { if (r) { if ("string" == typeof r) return _arrayLikeToArray(r, a); var t = {}.toString.call(r).slice(8, -1); return "Object" === t && r.constructor && (t = r.constructor.name), "Map" === t || "Set" === t ? Array.from(r) : "Arguments" === t || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(t) ? _arrayLikeToArray(r, a) : void 0; } }
-function _arrayLikeToArray(r, a) { (null == a || a > r.length) && (a = r.length); for (var e = 0, n = Array(a); e < a; e++) n[e] = r[e]; return n; }
+function file_loader_slicedToArray(r, e) { return file_loader_arrayWithHoles(r) || file_loader_iterableToArrayLimit(r, e) || file_loader_unsupportedIterableToArray(r, e) || file_loader_nonIterableRest(); }
+function file_loader_nonIterableRest() { throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }
+function file_loader_iterableToArrayLimit(r, l) { var t = null == r ? null : "undefined" != typeof Symbol && r[Symbol.iterator] || r["@@iterator"]; if (null != t) { var e, n, i, u, a = [], f = !0, o = !1; try { if (i = (t = t.call(r)).next, 0 === l) { if (Object(t) !== t) return; f = !1; } else for (; !(f = (e = i.call(t)).done) && (a.push(e.value), a.length !== l); f = !0); } catch (r) { o = !0, n = r; } finally { try { if (!f && null != t.return && (u = t.return(), Object(u) !== u)) return; } finally { if (o) throw n; } } return a; } }
+function file_loader_arrayWithHoles(r) { if (Array.isArray(r)) return r; }
+function _createForOfIteratorHelper(r, e) { var t = "undefined" != typeof Symbol && r[Symbol.iterator] || r["@@iterator"]; if (!t) { if (Array.isArray(r) || (t = file_loader_unsupportedIterableToArray(r)) || e && r && "number" == typeof r.length) { t && (r = t); var _n = 0, F = function F() {}; return { s: F, n: function n() { return _n >= r.length ? { done: !0 } : { done: !1, value: r[_n++] }; }, e: function e(r) { throw r; }, f: F }; } throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); } var o, a = !0, u = !1; return { s: function s() { t = t.call(r); }, n: function n() { var r = t.next(); return a = r.done, r; }, e: function e(r) { u = !0, o = r; }, f: function f() { try { a || null == t.return || t.return(); } finally { if (u) throw o; } } }; }
+function file_loader_unsupportedIterableToArray(r, a) { if (r) { if ("string" == typeof r) return file_loader_arrayLikeToArray(r, a); var t = {}.toString.call(r).slice(8, -1); return "Object" === t && r.constructor && (t = r.constructor.name), "Map" === t || "Set" === t ? Array.from(r) : "Arguments" === t || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(t) ? file_loader_arrayLikeToArray(r, a) : void 0; } }
+function file_loader_arrayLikeToArray(r, a) { (null == a || a > r.length) && (a = r.length); for (var e = 0, n = Array(a); e < a; e++) n[e] = r[e]; return n; }
 function _classCallCheck(a, n) { if (!(a instanceof n)) throw new TypeError("Cannot call a class as a function"); }
 function _defineProperties(e, r) { for (var t = 0; t < r.length; t++) { var o = r[t]; o.enumerable = o.enumerable || !1, o.configurable = !0, "value" in o && (o.writable = !0), Object.defineProperty(e, file_loader_toPropertyKey(o.key), o); } }
 function _createClass(e, r, t) { return r && _defineProperties(e.prototype, r), t && _defineProperties(e, t), Object.defineProperty(e, "prototype", { writable: !1 }), e; }
@@ -773,7 +809,7 @@ var PyodideFileLoader = /*#__PURE__*/function () {
       _step;
     try {
       for (_iterator.s(); !(_step = _iterator.n()).done;) {
-        var _step$value = _slicedToArray(_step.value, 2),
+        var _step$value = file_loader_slicedToArray(_step.value, 2),
           index = _step$value[0],
           file = _step$value[1];
         if (!file || file_loader_typeof(file) !== 'object') {