+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 
 
 
+# v0.0.2
+
+- Fix missive handling (ie parse at last moment, outside of the manager)
+- Fix `/scenery` 
+- Fix `/tests` (ie Flask cross origin)
+
+
 # v0.0.1 
 
 - Initial release